refactor(route): convert course-admin feedback route to TypeScript
Rewrite the course-admin/feedback route in TypeScript to improve type
safety and maintainability. Add explicit type annotations for services,
models, and the returned data structure. Replace JavaScript code with
modern TypeScript features, including non-null assertion and type
casting, to ensure better developer experience and reduce runtime errors.
