--- app/routes/course-admin/insights.js
@@ -1,20 +1,0 @@
-import { service } from '@ember/service';
-import BaseRoute from 'codecrafters-frontend/utils/base-route';
-
-export default class CourseInsightsRoute extends BaseRoute {
-  @service authenticator;
-  @service store;
-
-  async model() {
-    const course = this.modelFor('course-admin').course;
-
-    const courseInsightsDashboard = await this.store.queryRecord('course-insights-dashboard', {
-      course_id: course.id,
-      include: 'course',
-    });
-
-    return {
-      courseInsightsDashboard: courseInsightsDashboard,
-    };
-  }
-}
